requirements.txt
asgiref==3.6.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
djangorestframework==3.14.0
et-xmlfile==1.1.0
joblib==1.2.0
nltk==3.8.1
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
six==1.16.0
sqlparse==0.4.3
tqdm==4.65.0
tzdata==2022.7
